requirements.txt
prettytable
Pillow
requests